kitchen_factory/requirements.txt

Flask
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
Werkzeug
reportlab
arabic-reshaper
python-bidi
schedule
